requirements.txt
firebase-admin==6.*
functions-framework==3.*
google-cloud-aiplatform==1.*
google-cloud-bigquery==3.*
google-cloud-compute==1.*
google-cloud-pubsub==2.*
google-cloud-storage==2.*
openpyxl==3.*
pandas==2.*
vertexai==1.*
google-auth==2.*

flask==3.*
gunicorn==20.*
Werkzeug==3.*